Fixed a bunch of bugs
Shellshocks bullets can now hit the player, but not other enemies. Enemy detection is only triggered when an enemy is hit, not the player. New layers for enemies and enemy bullets. Moved the players collider to the parent object to make bullet hit detection work. Changed shellshocks to always shoot when the shoot animation is played, still not great as he will try to shoot through walls but better than playing the shoot animation without shooting. Updated prefabs.